chore: move more objects
