+++ src/Domain/Purchase/Subscription/ItemBasedRefund.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\ValueObjects;
+namespace Imdhemy\GooglePlay\Domain\Purchase\Subscription;
 
 /**
  * Used when a specific item should be refunded in a subscription with add-on items.
