+++ src/Domain/Purchase/Subscription/ProratedRefund.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\ValueObjects;
+namespace Imdhemy\GooglePlay\Domain\Purchase\Subscription;
 
 /**
  * Used when users should be refunded a prorated amount they paid for their subscription based on the amount of time
