+++ tests/Domain/Purchase/Subscription/FullRefundTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\ValueObjects;
+namespace Tests\Domain\Purchase\Subscription;
 
-use Imdhemy\GooglePlay\ValueObjects\FullRefund;
+use Imdhemy\GooglePlay\Domain\Purchase\Subscription\FullRefund;
 use Tests\TestCase;
 
 final class FullRefundTest extends TestCase
