+++ tests/Domain/Purchase/Subscription/ProratedRefundTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\ValueObjects;
+namespace Tests\Domain\Purchase\Subscription;
 
-use Imdhemy\GooglePlay\ValueObjects\ProratedRefund;
+use Imdhemy\GooglePlay\Domain\Purchase\Subscription\ProratedRefund;
 use Tests\TestCase;
 
 final class ProratedRefundTest extends TestCase
